[#15]refactor: footer -> copyright
